Add logic around _load_index_protobuf
