+++ automata/symbol/graph/graph_builder.py
@@ -5,7 +5,7 @@ Contains the `GraphBuilder` class, which builds a `SymbolGraph` from a correspon
 import logging
 import os
 import pickle
-from typing import Any
+from typing import Any, Optional
 
 import networkx as nx
 
@@ -20,12 +20,22 @@ from automata.symbol.symbol_utils import load_data_path
 logger = logging.getLogger(__name__)
 
 
+def _load_index_protobuf(path: str) -> Index:
+    """
+    Loads and returns an Index protobuf object from the given file path.
+    """
+    index = Index()
+    with open(path, "rb") as f:
+        index.ParseFromString(f.read())
+    return index
+
+
 class GraphBuilder:
     """Builds a `SymbolGraph` from a corresponding Index."""
 
     def __init__(
         self,
-        index: Index,
+        index_path: str,
         build_references: bool,
         build_relationships: bool,
         build_caller_relationships: bool,
@@ -33,7 +43,7 @@ class GraphBuilder:
         """
         Initializes a new instance of `GraphBuilder`.
         """
-        self.index = index
+        self.index_path = index_path
         self.build_references = build_references
         self.build_relationships = build_relationships
         self.build_caller_relationships = build_caller_relationships
@@ -57,14 +67,18 @@ class GraphBuilder:
         )
 
         if not from_pickle or not os.path.exists(graph_pickle_path):
-            for document in self.index.documents:
-                self._add_symbol_vertices(document)
-                if self.build_relationships:
-                    self._process_relationships(document)
-                if self.build_references:
-                    self._process_references(document)
-                if self.build_caller_relationships:
-                    self._process_caller_callee_relationships(document)
+            self.index = (
+                None if from_pickle else _load_index_protobuf(self.index_path)
+            )
+            if self.index is not None:
+                for document in self.index.documents:
+                    self._add_symbol_vertices(document)
+                    if self.build_relationships:
+                        self._process_relationships(document)
+                    if self.build_references:
+                        self._process_references(document)
+                    if self.build_caller_relationships:
+                        self._process_caller_callee_relationships(document)
 
             if save_graph_pickle:
                 with open(graph_pickle_path, "wb") as f:
